% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Medcode_atc_utils.R
\name{atc_convert}
\alias{atc_convert}
\title{Truncate an ATC Code to a Specified Level}
\usage{
atc_convert(code, level = 5)
}
\arguments{
\item{code}{A character scalar or vector of ATC codes, e.g. \code{"A01AA01"}.}

\item{level}{An integer scalar between 1 and 7 indicating how many characters to retain.
Defaults to \code{5}.}
}
\value{
A character vector of truncated ATC codes.
}
\description{
Truncates one or more ATC codes
to the given number of characters,
corresponding to
the desired ATC classification level
(e.g., anatomical main group, therapeutic
subgroup, chemical substance).
}
\examples{
atc_convert("A01AA01", level = 3)
atc_convert(c("B02BA02", "C03CA01"), level = 5)

}
