% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_CNNModel.R
\name{CNNBlock}
\alias{CNNBlock}
\title{CNNBlock Class}
\usage{
CNNBlock(in_channels, out_channels, spatial_dim)
}
\arguments{
\item{in_channels}{Number of input channels}

\item{out_channels}{Number of output channels}

\item{spatial_dim}{Spatial dimensionality (1, 2, or 3)}
}
\description{
Convolutional neural network block with residual connection.
}
\details{
Implements a residual CNN block with two convolutional layers,
batch normalization, and ReLU activation. Supports 1D, 2D, and 3D convolutions.
}
\keyword{internal}
