% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_CNNModel.R
\name{CNNLayer}
\alias{CNNLayer}
\title{CNNLayer Class}
\usage{
CNNLayer(input_size, hidden_size, num_layers = 1, spatial_dim = 1)
}
\arguments{
\item{input_size}{Number of input channels}

\item{hidden_size}{Number of hidden channels}

\item{num_layers}{Number of CNN blocks. Default 1}

\item{spatial_dim}{Spatial dimensionality (1, 2, or 3). Default 1}
}
\description{
Stack of CNN blocks with adaptive pooling.
}
\details{
Stacks multiple CNN blocks and applies adaptive average pooling
at the end. Supports 1D, 2D, and 3D spatial dimensions.
}
\keyword{internal}
