% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_utils.R
\name{collate_fn_dict_with_padding}
\alias{collate_fn_dict_with_padding}
\title{Collate Function with Padding}
\usage{
collate_fn_dict_with_padding(batch)
}
\arguments{
\item{batch}{A list of named lists, each representing a sample.}
}
\value{
A named list with padded tensors or lists.
}
\description{
Collates a batch of samples (list of named lists) into a single list with padded tensors.
}
