% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_ConCareModel.R
\name{ConCareLayerNorm}
\alias{ConCareLayerNorm}
\title{ConCare LayerNorm Class}
\usage{
ConCareLayerNorm(features, eps = 1e-07)
}
\arguments{
\item{features}{Number of features}

\item{eps}{Small constant for numerical stability. Default 1e-7}
}
\description{
Layer normalization.
}
\keyword{internal}
