% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_ConCareModel.R
\name{ConCareMultiHeadedAttention}
\alias{ConCareMultiHeadedAttention}
\title{ConCare MultiHeadedAttention Class}
\usage{
ConCareMultiHeadedAttention(h, d_model, dropout = 0)
}
\arguments{
\item{h}{Number of attention heads}

\item{d_model}{Model dimensionality}

\item{dropout}{Dropout rate. Default 0}
}
\description{
Multi-headed attention with DeCov regularization.
}
\keyword{internal}
