% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_ConCareModel.R
\name{ConCarePositionwiseFeedForward}
\alias{ConCarePositionwiseFeedForward}
\title{ConCare PositionwiseFeedForward Class}
\usage{
ConCarePositionwiseFeedForward(d_model, d_ff, dropout = 0.1)
}
\arguments{
\item{d_model}{Input/output dimensionality}

\item{d_ff}{Hidden dimensionality}

\item{dropout}{Dropout rate. Default 0.1}
}
\description{
Position-wise feed-forward network.
}
\keyword{internal}
