% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_ConCareModel.R
\name{ConCareSublayerConnection}
\alias{ConCareSublayerConnection}
\title{ConCare SublayerConnection Class}
\usage{
ConCareSublayerConnection(size, dropout)
}
\arguments{
\item{size}{Feature size}

\item{dropout}{Dropout rate}
}
\description{
Residual connection with layer norm.
}
\keyword{internal}
