% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trainer.R
\name{create_directory}
\alias{create_directory}
\title{Create Directory if Not Exists}
\usage{
create_directory(directory)
}
\arguments{
\item{directory}{Character. Path to directory.}
}
\value{
None.
}
\description{
Creates a directory recursively if it doesn't exist.
}
