% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_config.R
\name{DatasetConfig}
\alias{DatasetConfig}
\title{DatasetConfig: Root dataset configuration}
\description{
Describes the full dataset configuration including version and all table definitions.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{version}}{Version string of the dataset}

\item{\code{tables}}{Named list of TableConfig objects}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DatasetConfig-new}{\code{DatasetConfig$new()}}
\item \href{#method-DatasetConfig-validate}{\code{DatasetConfig$validate()}}
\item \href{#method-DatasetConfig-clone}{\code{DatasetConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetConfig-new"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetConfig-new}{}}}
\subsection{Method \code{new()}}{
Create a new DatasetConfig instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetConfig$new(version, tables)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{version}}{Dataset version string}

\item{\code{tables}}{Named list of TableConfig instances or raw config lists}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetConfig-validate"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetConfig-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the DatasetConfig object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetConfig$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetConfig-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetConfig-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
