% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processor_BaseProcessor.R
\name{DatasetProcessor}
\alias{DatasetProcessor}
\title{DatasetProcessor: Processor applied to entire datasets}
\description{
Optional class for processing entire datasets in bulk (e.g., batch statistics).
}
\section{Super class}{
\code{RHealth::Processor} -> \code{DatasetProcessor}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DatasetProcessor-process}{\code{DatasetProcessor$process()}}
\item \href{#method-DatasetProcessor-clone}{\code{DatasetProcessor$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="RHealth" data-topic="Processor" data-id="load"><a href='../../RHealth/html/Processor.html#method-Processor-load'><code>RHealth::Processor$load()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RHealth" data-topic="Processor" data-id="save"><a href='../../RHealth/html/Processor.html#method-Processor-save'><code>RHealth::Processor$save()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetProcessor-process"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetProcessor-process}{}}}
\subsection{Method \code{process()}}{
Abstract method: Process the entire dataset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetProcessor$process(samples)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{samples}}{A list of named lists representing all samples.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A processed list of named lists.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetProcessor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetProcessor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetProcessor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
