% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_BaseDataset.R
\name{.clean_path}
\alias{.clean_path}
\title{Normalizes a local file path or returns the URL unchanged.}
\usage{
.clean_path(path)
}
\arguments{
\item{path}{A character path or URL.}
}
\value{
Normalized character string.
}
\description{
Normalizes a local file path or returns the URL unchanged.
}
\keyword{internal}
