% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_BaseDataset.R
\name{.csv2parquet_path}
\alias{.csv2parquet_path}
\title{Given a *.csv(.gz) path, return *.parquet path in a /subset folder}
\usage{
.csv2parquet_path(csv_path)
}
\description{
Given a *.csv(.gz) path, return *.parquet path in a /subset folder
}
