% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_BaseDataset.R
\name{.find_path_with_fallback}
\alias{.find_path_with_fallback}
\title{Find an existing data file path with fallback for .gz extension.}
\usage{
.find_path_with_fallback(path)
}
\arguments{
\item{path}{A character path to a .csv, .csv.gz, .tsv, or .tsv.gz file.}
}
\value{
A list with \code{path} to an existing file and \code{separator} (',' or a tab).
}
\description{
This function checks for the existence of a path and its alternative with/without
\code{.gz}. It also determines the separator based on the file extension (.csv or .tsv).
}
\keyword{internal}
