% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_BaseDataset.R
\name{.is_url}
\alias{.is_url}
\title{Determines whether a path is an HTTP(S) URL.}
\usage{
.is_url(path)
}
\arguments{
\item{path}{A character string.}
}
\value{
Logical scalar indicating if it's a valid URL.
}
\description{
Determines whether a path is an HTTP(S) URL.
}
\keyword{internal}
