% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_BaseDataset.R
\name{.path_exists}
\alias{.path_exists}
\title{Tests whether a local path or remote URL exists.}
\usage{
.path_exists(path)
}
\arguments{
\item{path}{A character path or URL.}
}
\value{
Logical scalar indicating existence.
}
\description{
Tests whether a local path or remote URL exists.
}
\keyword{internal}
