% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_BaseDataset.R
\name{.scan_csv_gz_or_csv}
\alias{.scan_csv_gz_or_csv}
\title{Lazily loads a \code{.csv} or \code{.csv.gz} file and returns a Polars LazyFrame.
Automatically tries the alternate extension if the primary path fails.}
\usage{
.scan_csv_gz_or_csv(path)
}
\arguments{
\item{path}{File path or URL ending in \code{.csv} or \code{.csv.gz}.}
}
\value{
A polars LazyFrame.
}
\description{
Lazily loads a \code{.csv} or \code{.csv.gz} file and returns a Polars LazyFrame.
Automatically tries the alternate extension if the primary path fails.
}
\keyword{internal}
