% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_data.R
\name{Event-from_list}
\alias{Event-from_list}
\title{from_list: Create Event from row}
\arguments{
\item{row}{Named list of attributes.}

\item{event_type_col}{Name of column to use as event_type.}

\item{timestamp_col}{Name of column to use as timestamp.}
}
\value{
An Event R6 object.
}
\description{
Creates a new Event object from a named list.
}
\seealso{
\code{\link{Event}}
}
