% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processor_BaseProcessor.R
\name{FeatureProcessor}
\alias{FeatureProcessor}
\title{FeatureProcessor: Base class for all data processors}
\description{
Abstract class for input/output processors used in \code{rhealth}. Subclass this
to define how raw values (e.g. timestamps, labels) are transformed into
model-ready tensors.
}
\section{Methods to override}{

\itemize{
\item \code{process(value)}: Converts input value into tensor
\item \code{size()}: Returns output dimensionality
}
}

\section{Super class}{
\code{\link[RHealth:Processor]{RHealth::Processor}} -> \code{FeatureProcessor}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FeatureProcessor-fit}{\code{FeatureProcessor$fit()}}
\item \href{#method-FeatureProcessor-process}{\code{FeatureProcessor$process()}}
\item \href{#method-FeatureProcessor-clone}{\code{FeatureProcessor$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="RHealth" data-topic="Processor" data-id="load"><a href='../../RHealth/html/Processor.html#method-Processor-load'><code>RHealth::Processor$load()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RHealth" data-topic="Processor" data-id="save"><a href='../../RHealth/html/Processor.html#method-Processor-save'><code>RHealth::Processor$save()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureProcessor-fit"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureProcessor-fit}{}}}
\subsection{Method \code{fit()}}{
Fit processor using field values across samples.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureProcessor$fit(samples, field)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{samples}}{A list of named lists representing sample records.}

\item{\code{field}}{A string giving the field name to fit on.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureProcessor-process"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureProcessor-process}{}}}
\subsection{Method \code{process()}}{
Abstract method: Process an individual field value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureProcessor$process(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A raw value (e.g., character, number, etc).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A processed value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureProcessor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureProcessor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureProcessor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
