% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_ConCareModel.R
\name{FinalAttentionQKV}
\alias{FinalAttentionQKV}
\title{FinalAttentionQKV Class}
\usage{
FinalAttentionQKV(
  attention_input_dim,
  attention_hidden_dim,
  attention_type = "add",
  dropout = 0.5
)
}
\arguments{
\item{attention_input_dim}{Input dimensionality}

\item{attention_hidden_dim}{Hidden dimensionality}

\item{attention_type}{Type of attention ("add", "mul", "concat"). Default "add"}

\item{dropout}{Dropout rate. Default 0.5}
}
\description{
Final attention layer using query, key, value mechanism.
}
\details{
Computes attention weights for the final aggregation of temporal representations.
}
\keyword{internal}
