% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_utils.R
\name{get_dataloader}
\alias{get_dataloader}
\title{Create DataLoader}
\usage{
get_dataloader(dataset, batch_size, shuffle = FALSE)
}
\arguments{
\item{dataset}{A torch dataset or dataset_subset object.}

\item{batch_size}{Integer, number of samples per batch.}

\item{shuffle}{Logical, whether to shuffle the data. Default: FALSE.}
}
\value{
A torch dataloader object.
}
\description{
Creates a torch dataloader with padding-aware collate function.
}
