% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Medcode_inner_map.R
\name{get_descendants}
\alias{get_descendants}
\title{Get Descendant Codes in the Hierarchy}
\usage{
get_descendants(code, system = "ICD9CM")
}
\arguments{
\item{code}{A single \code{character} string specifying the starting code.}

\item{system}{A single \code{character} string naming the code system, one of
\code{supported_inner()}. Defaults to \code{"ICD9CM"}.}
}
\value{
A character vector of all descendant codes, in no particular order.
}
\description{
For a given code,
returns all descendant codes
(children, grandchildren, etc.)
by traversing the hierarchy downward.
}
\examples{
\dontrun{
get_descendants("A00", "ICD10CM")
}

}
\seealso{
\code{\link{get_ancestors}}, \code{\link{load_medcode}}
}
