% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trainer.R
\name{get_metrics_fn}
\alias{get_metrics_fn}
\title{Get Metrics Function}
\usage{
get_metrics_fn(mode)
}
\arguments{
\item{mode}{Character. One of "binary", "multiclass", "multilabel", or "regression".}
}
\value{
Function. Metrics calculation function.
}
\description{
Returns appropriate metric function according to task mode.
}
