% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Task_InHospitalMortalityMIMIC4.R
\name{InHospitalMortalityMIMIC4}
\alias{InHospitalMortalityMIMIC4}
\title{InHospitalMortalityMIMIC4 Task}
\description{
Task for predicting in-hospital mortality using MIMIC-IV dataset.
This task leverages lab results from the first 48 hours of an admission to
predict the likelihood of in-hospital mortality.
}
\section{Super class}{
\code{\link[RHealth:BaseTask]{RHealth::BaseTask}} -> \code{InHospitalMortalityMIMIC4}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{task_name}}{The name of the task.}

\item{\code{input_schema}}{The schema for input data.}

\item{\code{output_schema}}{The schema for output data.}

\item{\code{label}}{The name of the label column.}

\item{\code{LABITEMS}}{A list of lab item IDs used in this task.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InHospitalMortalityMIMIC4-new}{\code{InHospitalMortalityMIMIC4$new()}}
\item \href{#method-InHospitalMortalityMIMIC4-pre_filter}{\code{InHospitalMortalityMIMIC4$pre_filter()}}
\item \href{#method-InHospitalMortalityMIMIC4-call}{\code{InHospitalMortalityMIMIC4$call()}}
\item \href{#method-InHospitalMortalityMIMIC4-clone}{\code{InHospitalMortalityMIMIC4$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InHospitalMortalityMIMIC4-new"></a>}}
\if{latex}{\out{\hypertarget{method-InHospitalMortalityMIMIC4-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new InHospitalMortalityMIMIC4 instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InHospitalMortalityMIMIC4$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InHospitalMortalityMIMIC4-pre_filter"></a>}}
\if{latex}{\out{\hypertarget{method-InHospitalMortalityMIMIC4-pre_filter}{}}}
\subsection{Method \code{pre_filter()}}{
Pre-filter hook to retain only necessary columns for this task.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InHospitalMortalityMIMIC4$pre_filter(df)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{A lazy query containing all events.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A filtered LazyFrame with only relevant columns.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InHospitalMortalityMIMIC4-call"></a>}}
\if{latex}{\out{\hypertarget{method-InHospitalMortalityMIMIC4-call}{}}}
\subsection{Method \code{call()}}{
Main processing method to generate samples.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InHospitalMortalityMIMIC4$call(patient)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{patient}}{An object with method \code{get_events(event_type, ...)}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of samples.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InHospitalMortalityMIMIC4-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InHospitalMortalityMIMIC4-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InHospitalMortalityMIMIC4$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
