% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_SampleDataset.R
\name{load_sample_dataset}
\alias{load_sample_dataset}
\title{Load a SampleDataset object from a directory}
\usage{
load_sample_dataset(path)
}
\arguments{
\item{path}{The directory path from where to load the dataset.}
}
\value{
The reconstructed SampleDataset object.
}
\description{
This function reconstructs a SampleDataset object from a directory.
}
