% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_MIMIC4EHRDataset.R
\name{log_memory_usage}
\alias{log_memory_usage}
\title{Logs memory usage (in MB) using ps::ps_memory_info if available.}
\usage{
log_memory_usage(tag = "")
}
\arguments{
\item{tag}{Optional label for the logging context.}
}
\description{
Logs memory usage (in MB) using ps::ps_memory_info if available.
}
\keyword{internal}
