% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_MIMIC3Dataset.R
\docType{class}
\name{MIMIC3Dataset}
\alias{MIMIC3Dataset}
\title{MIMIC3Dataset: Dataset class for MIMIC-III}
\usage{
\method{MIMIC3Dataset}{initialize}(...)
}
\description{
MIMIC3Dataset: Dataset class for MIMIC-III

MIMIC3Dataset: Dataset class for MIMIC-III
}
\details{
This class inherits from BaseDataset and is specialized for handling MIMIC-III data.
It ensures key tables like patients, admissions, and icustays are loaded,
and allows appending additional tables. Also provides per-table preprocessing.
}
\section{Super class}{
\code{RHealth::BaseDataset} -> \code{MIMIC3Dataset}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MIMIC3Dataset-new}{\code{MIMIC3Dataset$new()}}
\item \href{#method-MIMIC3Dataset-preprocess_noteevents}{\code{MIMIC3Dataset$preprocess_noteevents()}}
\item \href{#method-MIMIC3Dataset-clone}{\code{MIMIC3Dataset$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="RHealth" data-topic="BaseDataset" data-id="collected_global_event_df"><a href='../../RHealth/html/BaseDataset.html#method-BaseDataset-collected_global_event_df'><code>RHealth::BaseDataset$collected_global_event_df()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RHealth" data-topic="BaseDataset" data-id="default_task"><a href='../../RHealth/html/BaseDataset.html#method-BaseDataset-default_task'><code>RHealth::BaseDataset$default_task()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RHealth" data-topic="BaseDataset" data-id="get_patient"><a href='../../RHealth/html/BaseDataset.html#method-BaseDataset-get_patient'><code>RHealth::BaseDataset$get_patient()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RHealth" data-topic="BaseDataset" data-id="iter_patients"><a href='../../RHealth/html/BaseDataset.html#method-BaseDataset-iter_patients'><code>RHealth::BaseDataset$iter_patients()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RHealth" data-topic="BaseDataset" data-id="load_data"><a href='../../RHealth/html/BaseDataset.html#method-BaseDataset-load_data'><code>RHealth::BaseDataset$load_data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RHealth" data-topic="BaseDataset" data-id="load_table"><a href='../../RHealth/html/BaseDataset.html#method-BaseDataset-load_table'><code>RHealth::BaseDataset$load_table()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RHealth" data-topic="BaseDataset" data-id="set_task"><a href='../../RHealth/html/BaseDataset.html#method-BaseDataset-set_task'><code>RHealth::BaseDataset$set_task()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RHealth" data-topic="BaseDataset" data-id="stats"><a href='../../RHealth/html/BaseDataset.html#method-BaseDataset-stats'><code>RHealth::BaseDataset$stats()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RHealth" data-topic="BaseDataset" data-id="unique_patient_ids"><a href='../../RHealth/html/BaseDataset.html#method-BaseDataset-unique_patient_ids'><code>RHealth::BaseDataset$unique_patient_ids()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MIMIC3Dataset-new"></a>}}
\if{latex}{\out{\hypertarget{method-MIMIC3Dataset-new}{}}}
\subsection{Method \code{new()}}{
initialize MIMIC3Dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MIMIC3Dataset$new(
  root,
  tables = character(),
  dataset_name = NULL,
  config_path = NULL,
  dev = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{root}}{Root directory of the dataset.}

\item{\code{tables}}{Character vector of extra tables to include.}

\item{\code{dataset_name}}{Optional dataset name.}

\item{\code{config_path}}{Optional path to YAML config file.}

\item{\code{dev}}{Logical flag for dev mode.}

\item{\code{...}}{Additional arguments passed to \code{BaseDataset}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MIMIC3Dataset-preprocess_noteevents"></a>}}
\if{latex}{\out{\hypertarget{method-MIMIC3Dataset-preprocess_noteevents}{}}}
\subsection{Method \code{preprocess_noteevents()}}{
Table-specific preprocessing for noteevents.
If \code{charttime} is missing, fills it with \code{chartdate} + " 00:00:00".
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MIMIC3Dataset$preprocess_noteevents(df)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{A polars LazyFrame.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A modified LazyFrame.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MIMIC3Dataset-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MIMIC3Dataset-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MIMIC3Dataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
