% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Task_MortalityPredictionOMOP.R
\name{MortalityPredictionOMOP}
\alias{MortalityPredictionOMOP}
\title{MortalityPredictionOMOP Task}
\description{
Task for predicting mortality using OMOP CDM dataset.
Predicts whether a patient has a death record based on clinical information from each visit.
}
\section{Super class}{
\code{\link[RHealth:BaseTask]{RHealth::BaseTask}} -> \code{MortalityPredictionOMOP}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{task_name}}{Name of the task.}

\item{\code{input_schema}}{Input schema.}

\item{\code{output_schema}}{Output schema.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MortalityPredictionOMOP-new}{\code{MortalityPredictionOMOP$new()}}
\item \href{#method-MortalityPredictionOMOP-pre_filter}{\code{MortalityPredictionOMOP$pre_filter()}}
\item \href{#method-MortalityPredictionOMOP-call}{\code{MortalityPredictionOMOP$call()}}
\item \href{#method-MortalityPredictionOMOP-clone}{\code{MortalityPredictionOMOP$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MortalityPredictionOMOP-new"></a>}}
\if{latex}{\out{\hypertarget{method-MortalityPredictionOMOP-new}{}}}
\subsection{Method \code{new()}}{
Initialize the task.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MortalityPredictionOMOP$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MortalityPredictionOMOP-pre_filter"></a>}}
\if{latex}{\out{\hypertarget{method-MortalityPredictionOMOP-pre_filter}{}}}
\subsection{Method \code{pre_filter()}}{
Pre-filter hook to retain only necessary columns.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MortalityPredictionOMOP$pre_filter(df)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{A lazy query containing all events.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A filtered LazyFrame.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MortalityPredictionOMOP-call"></a>}}
\if{latex}{\out{\hypertarget{method-MortalityPredictionOMOP-call}{}}}
\subsection{Method \code{call()}}{
Process a single patient to generate samples.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MortalityPredictionOMOP$call(patient)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{patient}}{A Patient object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of samples.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MortalityPredictionOMOP-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MortalityPredictionOMOP-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MortalityPredictionOMOP$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
