% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processor_multiclass.R
\name{MultiClassLabelProcessor}
\alias{MultiClassLabelProcessor}
\title{Multi-Class Label Processor}
\description{
Processor for multi-class classification tasks. Converts string or integer labels
into integer indices, with one output per label.
}
\section{Super classes}{
\code{\link[RHealth:Processor]{RHealth::Processor}} -> \code{\link[RHealth:FeatureProcessor]{RHealth::FeatureProcessor}} -> \code{MultiClassLabelProcessor}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{label_vocab}}{A named integer vector mapping labels to indices.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MultiClassLabelProcessor-new}{\code{MultiClassLabelProcessor$new()}}
\item \href{#method-MultiClassLabelProcessor-fit}{\code{MultiClassLabelProcessor$fit()}}
\item \href{#method-MultiClassLabelProcessor-process}{\code{MultiClassLabelProcessor$process()}}
\item \href{#method-MultiClassLabelProcessor-size}{\code{MultiClassLabelProcessor$size()}}
\item \href{#method-MultiClassLabelProcessor-print}{\code{MultiClassLabelProcessor$print()}}
\item \href{#method-MultiClassLabelProcessor-clone}{\code{MultiClassLabelProcessor$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="RHealth" data-topic="Processor" data-id="load"><a href='../../RHealth/html/Processor.html#method-Processor-load'><code>RHealth::Processor$load()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RHealth" data-topic="Processor" data-id="save"><a href='../../RHealth/html/Processor.html#method-Processor-save'><code>RHealth::Processor$save()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiClassLabelProcessor-new"></a>}}
\if{latex}{\out{\hypertarget{method-MultiClassLabelProcessor-new}{}}}
\subsection{Method \code{new()}}{
Initialize with empty label vocabulary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiClassLabelProcessor$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiClassLabelProcessor-fit"></a>}}
\if{latex}{\out{\hypertarget{method-MultiClassLabelProcessor-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the label vocabulary from the given field of all samples.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiClassLabelProcessor$fit(samples, field)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{samples}}{A list of named lists representing the dataset.}

\item{\code{field}}{The name of the field containing labels.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiClassLabelProcessor-process"></a>}}
\if{latex}{\out{\hypertarget{method-MultiClassLabelProcessor-process}{}}}
\subsection{Method \code{process()}}{
Convert a label into a torch long integer tensor (scalar).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiClassLabelProcessor$process(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The raw label value.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An int64 torch tensor.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiClassLabelProcessor-size"></a>}}
\if{latex}{\out{\hypertarget{method-MultiClassLabelProcessor-size}{}}}
\subsection{Method \code{size()}}{
Return number of classes (vocabulary size).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiClassLabelProcessor$size()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiClassLabelProcessor-print"></a>}}
\if{latex}{\out{\hypertarget{method-MultiClassLabelProcessor-print}{}}}
\subsection{Method \code{print()}}{
Print a summary of the processor.


Print a summary of the processor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiClassLabelProcessor$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiClassLabelProcessor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MultiClassLabelProcessor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiClassLabelProcessor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
