% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processor_multilabel.R
\name{MultiLabelProcessor}
\alias{MultiLabelProcessor}
\title{Multi-Label Processor}
\description{
Processor for multi-label classification. Converts a list of active labels into a
one-hot tensor with multiple 1s. Inherits from FeatureProcessor.
}
\section{Super classes}{
\code{\link[RHealth:Processor]{RHealth::Processor}} -> \code{\link[RHealth:FeatureProcessor]{RHealth::FeatureProcessor}} -> \code{MultiLabelProcessor}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{label_vocab}}{A named integer vector mapping label values to index positions.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MultiLabelProcessor-new}{\code{MultiLabelProcessor$new()}}
\item \href{#method-MultiLabelProcessor-fit}{\code{MultiLabelProcessor$fit()}}
\item \href{#method-MultiLabelProcessor-process}{\code{MultiLabelProcessor$process()}}
\item \href{#method-MultiLabelProcessor-size}{\code{MultiLabelProcessor$size()}}
\item \href{#method-MultiLabelProcessor-print}{\code{MultiLabelProcessor$print()}}
\item \href{#method-MultiLabelProcessor-clone}{\code{MultiLabelProcessor$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="RHealth" data-topic="Processor" data-id="load"><a href='../../RHealth/html/Processor.html#method-Processor-load'><code>RHealth::Processor$load()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RHealth" data-topic="Processor" data-id="save"><a href='../../RHealth/html/Processor.html#method-Processor-save'><code>RHealth::Processor$save()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiLabelProcessor-new"></a>}}
\if{latex}{\out{\hypertarget{method-MultiLabelProcessor-new}{}}}
\subsection{Method \code{new()}}{
Constructor that initializes the vocabulary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiLabelProcessor$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiLabelProcessor-fit"></a>}}
\if{latex}{\out{\hypertarget{method-MultiLabelProcessor-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the processor from all multi-label lists across samples.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiLabelProcessor$fit(samples, field)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{samples}}{A list of named lists (sample records).}

\item{\code{field}}{The name of the multi-label field.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiLabelProcessor-process"></a>}}
\if{latex}{\out{\hypertarget{method-MultiLabelProcessor-process}{}}}
\subsection{Method \code{process()}}{
Process a list of active labels into a one-hot float tensor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiLabelProcessor$process(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A character or numeric vector of active labels.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A torch tensor of shape \code{num_classes} with 0s and 1s.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiLabelProcessor-size"></a>}}
\if{latex}{\out{\hypertarget{method-MultiLabelProcessor-size}{}}}
\subsection{Method \code{size()}}{
Return number of classes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiLabelProcessor$size()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Integer.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiLabelProcessor-print"></a>}}
\if{latex}{\out{\hypertarget{method-MultiLabelProcessor-print}{}}}
\subsection{Method \code{print()}}{
Print a summary of the processor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiLabelProcessor$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiLabelProcessor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MultiLabelProcessor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiLabelProcessor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
