% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Task_NextMortalityMIMIC4.R
\name{NextMortalityMIMIC4}
\alias{NextMortalityMIMIC4}
\title{NextMortalityMIMIC4 Task}
\description{
Task for predicting in-hospital mortality using MIMIC-IV dataset.
Uses lab results from the first 48 hours after admission as input features.
}
\section{Super class}{
\code{\link[RHealth:BaseTask]{RHealth::BaseTask}} -> \code{NextMortalityMIMIC4}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{label}}{the name of the label column.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NextMortalityMIMIC4-new}{\code{NextMortalityMIMIC4$new()}}
\item \href{#method-NextMortalityMIMIC4-pre_filter}{\code{NextMortalityMIMIC4$pre_filter()}}
\item \href{#method-NextMortalityMIMIC4-call}{\code{NextMortalityMIMIC4$call()}}
\item \href{#method-NextMortalityMIMIC4-clone}{\code{NextMortalityMIMIC4$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NextMortalityMIMIC4-new"></a>}}
\if{latex}{\out{\hypertarget{method-NextMortalityMIMIC4-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new NextMortalityMIMIC4 instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NextMortalityMIMIC4$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NextMortalityMIMIC4-pre_filter"></a>}}
\if{latex}{\out{\hypertarget{method-NextMortalityMIMIC4-pre_filter}{}}}
\subsection{Method \code{pre_filter()}}{
Pre-filter hook to retain only necessary columns for this task.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NextMortalityMIMIC4$pre_filter(df)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{A lazy query containing all events.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A filtered LazyFrame with only relevant columns.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NextMortalityMIMIC4-call"></a>}}
\if{latex}{\out{\hypertarget{method-NextMortalityMIMIC4-call}{}}}
\subsection{Method \code{call()}}{
Main processing method to generate samples.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NextMortalityMIMIC4$call(patient)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{patient}}{An object with method \code{get_events(event_type, ...)}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of samples.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NextMortalityMIMIC4-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NextMortalityMIMIC4-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NextMortalityMIMIC4$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
