% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_TransformerModel.R
\name{PositionwiseFeedForward}
\alias{PositionwiseFeedForward}
\title{PositionwiseFeedForward Class}
\usage{
PositionwiseFeedForward(d_model, d_ff, dropout = 0.1)
}
\arguments{
\item{d_model}{Input/output dimensionality}

\item{d_ff}{Hidden layer dimensionality}

\item{dropout}{Dropout rate}
}
\description{
Two-layer feed-forward network with GELU activation.
}
\keyword{internal}
