% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processor_raw.R
\name{RawProcessor}
\alias{RawProcessor}
\title{Raw Processor}
\description{
Processor that returns raw values without any transformation.
Inherits from FeatureProcessor.
}
\section{Super classes}{
\code{RHealth::Processor} -> \code{RHealth::FeatureProcessor} -> \code{RawProcessor}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RawProcessor-process}{\code{RawProcessor$process()}}
\item \href{#method-RawProcessor-size}{\code{RawProcessor$size()}}
\item \href{#method-RawProcessor-print}{\code{RawProcessor$print()}}
\item \href{#method-RawProcessor-clone}{\code{RawProcessor$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="RHealth" data-topic="Processor" data-id="load"><a href='../../RHealth/html/Processor.html#method-Processor-load'><code>RHealth::Processor$load()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RHealth" data-topic="Processor" data-id="save"><a href='../../RHealth/html/Processor.html#method-Processor-save'><code>RHealth::Processor$save()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RHealth" data-topic="FeatureProcessor" data-id="fit"><a href='../../RHealth/html/FeatureProcessor.html#method-FeatureProcessor-fit'><code>RHealth::FeatureProcessor$fit()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawProcessor-process"></a>}}
\if{latex}{\out{\hypertarget{method-RawProcessor-process}{}}}
\subsection{Method \code{process()}}{
Return the raw input as-is.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawProcessor$process(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A raw field value (any type).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The unmodified input.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawProcessor-size"></a>}}
\if{latex}{\out{\hypertarget{method-RawProcessor-size}{}}}
\subsection{Method \code{size()}}{
Optional: Return size/dimension of processed output.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawProcessor$size()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawProcessor-print"></a>}}
\if{latex}{\out{\hypertarget{method-RawProcessor-print}{}}}
\subsection{Method \code{print()}}{
Print a string representation of the processor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawProcessor$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawProcessor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RawProcessor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawProcessor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
