% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_AdaCareModel.R
\name{Recalibration}
\alias{Recalibration}
\title{Recalibration Class}
\usage{
Recalibration(channel, reduction = 9, activation = "sigmoid")
}
\arguments{
\item{channel}{Number of input channels}

\item{reduction}{Reduction ratio for bottleneck. Default 9}

\item{activation}{Activation function ("sigmoid", "sparsemax", "softmax"). Default "sigmoid"}
}
\description{
Feature recalibration module using squeeze-and-excitation mechanism.
}
\details{
Adaptively recalibrates channel-wise feature responses by explicitly modeling
interdependencies between channels.
}
\keyword{internal}
