% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processor_regression.R
\name{RegressionLabelProcessor}
\alias{RegressionLabelProcessor}
\title{Regression Label Processor}
\description{
Processor for scalar regression labels. Converts values to a 1D float tensor.
}
\section{Super classes}{
\code{RHealth::Processor} -> \code{RHealth::FeatureProcessor} -> \code{RegressionLabelProcessor}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RegressionLabelProcessor-process}{\code{RegressionLabelProcessor$process()}}
\item \href{#method-RegressionLabelProcessor-size}{\code{RegressionLabelProcessor$size()}}
\item \href{#method-RegressionLabelProcessor-print}{\code{RegressionLabelProcessor$print()}}
\item \href{#method-RegressionLabelProcessor-clone}{\code{RegressionLabelProcessor$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="RHealth" data-topic="Processor" data-id="load"><a href='../../RHealth/html/Processor.html#method-Processor-load'><code>RHealth::Processor$load()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RHealth" data-topic="Processor" data-id="save"><a href='../../RHealth/html/Processor.html#method-Processor-save'><code>RHealth::Processor$save()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RHealth" data-topic="FeatureProcessor" data-id="fit"><a href='../../RHealth/html/FeatureProcessor.html#method-FeatureProcessor-fit'><code>RHealth::FeatureProcessor$fit()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegressionLabelProcessor-process"></a>}}
\if{latex}{\out{\hypertarget{method-RegressionLabelProcessor-process}{}}}
\subsection{Method \code{process()}}{
Process a numeric label into a single-element float tensor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegressionLabelProcessor$process(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A numeric value.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A torch tensor of shape \link{1}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegressionLabelProcessor-size"></a>}}
\if{latex}{\out{\hypertarget{method-RegressionLabelProcessor-size}{}}}
\subsection{Method \code{size()}}{
Return the size of the processed label (always 1).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegressionLabelProcessor$size()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Integer 1
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegressionLabelProcessor-print"></a>}}
\if{latex}{\out{\hypertarget{method-RegressionLabelProcessor-print}{}}}
\subsection{Method \code{print()}}{
Print a string representation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegressionLabelProcessor$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegressionLabelProcessor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RegressionLabelProcessor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegressionLabelProcessor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
