% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHealth-package.R
\docType{package}
\name{RHealth-package}
\alias{RHealth}
\alias{RHealth-package}
\title{RHealth: A Deep Learning Toolkit for Healthcare Predictive Modeling}
\description{
RHealth is an open-source R package specifically designed to bring comprehensive deep learning toolkits to the R community for healthcare predictive modeling.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://v1xerunt.github.io/RHealth}
  \item \url{https://github.com/v1xerunt/RHealth}
  \item Report bugs at \url{https://github.com/v1xerunt/RHealth/issues}
}

}
\author{
\strong{Maintainer}: Junyi Gao \email{junyii.gao@gmail.com}

Authors:
\itemize{
  \item Ji Song \email{eaad0907@163.com}
  \item Zhixia Ren \email{jexica0921@gmail.com}
  \item Zhenbang Wu \email{zw12@illinois.edu}
  \item John Wu \email{johnwu3@illinois.edu}
  \item Chaoqi Yang \email{ycqsjtu@gmail.com}
  \item Jimeng Sun \email{jimeng@illinois.edu}
  \item Liantao Ma \email{malt@pku.edu.cn}
  \item Ewen Harrison \email{ewen.harrison@ed.ac.uk}
}

}
\keyword{internal}
