% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_config.R
\name{rhealth.config}
\alias{rhealth.config}
\title{Dataset Configuration}
\description{
This module defines R6 classes for representing configuration files for PyHealth datasets,
including table join specifications, per-table attributes, and the overall dataset structure.
It also provides a function to load configuration from a YAML file.
}
