% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processor_BaseProcessor.R
\name{SampleProcessor}
\alias{SampleProcessor}
\title{SampleProcessor: Processor for sample-level transformations}
\description{
Optional processor for transformations applied at the whole-sample level
(e.g., normalizing an image+label pair).
}
\section{Super class}{
\code{RHealth::Processor} -> \code{SampleProcessor}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SampleProcessor-process}{\code{SampleProcessor$process()}}
\item \href{#method-SampleProcessor-clone}{\code{SampleProcessor$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="RHealth" data-topic="Processor" data-id="load"><a href='../../RHealth/html/Processor.html#method-Processor-load'><code>RHealth::Processor$load()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RHealth" data-topic="Processor" data-id="save"><a href='../../RHealth/html/Processor.html#method-Processor-save'><code>RHealth::Processor$save()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SampleProcessor-process"></a>}}
\if{latex}{\out{\hypertarget{method-SampleProcessor-process}{}}}
\subsection{Method \code{process()}}{
Abstract method: Process a single sample (a named list).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SampleProcessor$process(sample)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sample}}{A named list representing one data sample.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A processed named list.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SampleProcessor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SampleProcessor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SampleProcessor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
