% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processor_sequence.R
\name{SequenceProcessor}
\alias{SequenceProcessor}
\title{Sequence Processor}
\description{
Feature processor for encoding categorical sequences (e.g., medical codes) into
numerical indices. Supports dynamic vocabulary construction.
}
\section{Super classes}{
\code{\link[RHealth:Processor]{RHealth::Processor}} -> \code{\link[RHealth:FeatureProcessor]{RHealth::FeatureProcessor}} -> \code{SequenceProcessor}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{code_vocab}}{A named integer vector representing token-to-index mappings.}

\item{\code{.next_index}}{The next available index for unseen tokens.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SequenceProcessor-new}{\code{SequenceProcessor$new()}}
\item \href{#method-SequenceProcessor-process}{\code{SequenceProcessor$process()}}
\item \href{#method-SequenceProcessor-size}{\code{SequenceProcessor$size()}}
\item \href{#method-SequenceProcessor-print}{\code{SequenceProcessor$print()}}
\item \href{#method-SequenceProcessor-clone}{\code{SequenceProcessor$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="RHealth" data-topic="Processor" data-id="load"><a href='../../RHealth/html/Processor.html#method-Processor-load'><code>RHealth::Processor$load()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RHealth" data-topic="Processor" data-id="save"><a href='../../RHealth/html/Processor.html#method-Processor-save'><code>RHealth::Processor$save()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RHealth" data-topic="FeatureProcessor" data-id="fit"><a href='../../RHealth/html/FeatureProcessor.html#method-FeatureProcessor-fit'><code>RHealth::FeatureProcessor$fit()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SequenceProcessor-new"></a>}}
\if{latex}{\out{\hypertarget{method-SequenceProcessor-new}{}}}
\subsection{Method \code{new()}}{
Initialize with default vocabulary for \if{html}{\out{<pad>}} and \if{html}{\out{<unk>}}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SequenceProcessor$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SequenceProcessor-process"></a>}}
\if{latex}{\out{\hypertarget{method-SequenceProcessor-process}{}}}
\subsection{Method \code{process()}}{
Process a sequence of tokens into a tensor of indices.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SequenceProcessor$process(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A character vector of tokens.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A long-type tensor of indices.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SequenceProcessor-size"></a>}}
\if{latex}{\out{\hypertarget{method-SequenceProcessor-size}{}}}
\subsection{Method \code{size()}}{
Return size of vocabulary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SequenceProcessor$size()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SequenceProcessor-print"></a>}}
\if{latex}{\out{\hypertarget{method-SequenceProcessor-print}{}}}
\subsection{Method \code{print()}}{
Print summary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SequenceProcessor$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SequenceProcessor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SequenceProcessor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SequenceProcessor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
