% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trainer.R
\name{set_logger}
\alias{set_logger}
\title{Initialize Logger}
\usage{
set_logger(log_path = NULL)
}
\arguments{
\item{log_path}{Character. Path to save log file.}
}
\value{
None
}
\description{
Sets up the logger to write into a specified file using \code{futile.logger}.
}
