% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_AdaCareModel.R
\name{Sparsemax}
\alias{Sparsemax}
\title{Sparsemax Class}
\usage{
Sparsemax(dim = NULL)
}
\arguments{
\item{dim}{Dimension along which to apply sparsemax. Default -1.}
}
\description{
Sparsemax activation function as an alternative to softmax.
}
\details{
Produces sparse probability distributions, which can be beneficial for
interpretability by setting low-importance weights to exactly zero.
}
\keyword{internal}
