% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_splitter.R
\name{split_by_sample}
\alias{split_by_sample}
\title{Dataset Split Functions}
\usage{
split_by_sample(
  dataset,
  ratios,
  seed = NULL,
  stratify = FALSE,
  stratify_by = NULL,
  get_index = FALSE
)
}
\arguments{
\item{dataset}{A \code{SampleDataset} object.}

\item{ratios}{A numeric vector of length 3 indicating train/val/test split ratios. Must sum to 1.}

\item{seed}{Optional integer for reproducibility.}

\item{stratify}{Logical, whether to perform stratified sampling. Default: FALSE.}

\item{stratify_by}{Character, the name of the field to stratify by (e.g., the label). Required if \code{stratify} is TRUE.}

\item{get_index}{Logical, whether to return the indices instead of subsets. Default: FALSE.}
}
\value{
A list of 3 torch::dataset_subset objects or 3 tensors of indices if get_index = TRUE.
}
\description{
Provides functions to split a \code{SampleDataset} object by sample, patient, or visit.
}
