% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_TransformerModel.R
\name{SublayerConnection}
\alias{SublayerConnection}
\title{SublayerConnection Class}
\usage{
SublayerConnection(size, dropout)
}
\arguments{
\item{size}{Feature dimensionality}

\item{dropout}{Dropout rate}
}
\description{
Pre-norm residual connection wrapper.
}
\keyword{internal}
