% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Medcode_inner_map.R
\name{supported_inner}
\alias{supported_inner}
\title{List Supported Medical Code Systems}
\usage{
supported_inner()
}
\value{
A character vector of supported code system identifiers.
}
\description{
Returns a character vector of the internal code systems currently supported
by the package.
}
\examples{
supported_inner()

}
\concept{MedCode}
