% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processor_text.R
\name{TextProcessor}
\alias{TextProcessor}
\title{Text Processor}
\description{
Processor for textual input. Inherits from FeatureProcessor and defines a minimal no-op process method.
}
\section{Super classes}{
\code{RHealth::Processor} -> \code{RHealth::FeatureProcessor} -> \code{TextProcessor}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TextProcessor-process}{\code{TextProcessor$process()}}
\item \href{#method-TextProcessor-size}{\code{TextProcessor$size()}}
\item \href{#method-TextProcessor-print}{\code{TextProcessor$print()}}
\item \href{#method-TextProcessor-clone}{\code{TextProcessor$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="RHealth" data-topic="Processor" data-id="load"><a href='../../RHealth/html/Processor.html#method-Processor-load'><code>RHealth::Processor$load()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RHealth" data-topic="Processor" data-id="save"><a href='../../RHealth/html/Processor.html#method-Processor-save'><code>RHealth::Processor$save()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RHealth" data-topic="FeatureProcessor" data-id="fit"><a href='../../RHealth/html/FeatureProcessor.html#method-FeatureProcessor-fit'><code>RHealth::FeatureProcessor$fit()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TextProcessor-process"></a>}}
\if{latex}{\out{\hypertarget{method-TextProcessor-process}{}}}
\subsection{Method \code{process()}}{
Process a raw text input. (Currently a no-op identity function.)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextProcessor$process(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A single text input (character string).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The processed text (same as input).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TextProcessor-size"></a>}}
\if{latex}{\out{\hypertarget{method-TextProcessor-size}{}}}
\subsection{Method \code{size()}}{
Optional: Return size or dimensionality if applicable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextProcessor$size()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL by default.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TextProcessor-print"></a>}}
\if{latex}{\out{\hypertarget{method-TextProcessor-print}{}}}
\subsection{Method \code{print()}}{
Return a printable string representation of the processor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextProcessor$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character string.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TextProcessor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TextProcessor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextProcessor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
