% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_TransformerModel.R
\name{TransformerBlock}
\alias{TransformerBlock}
\title{TransformerBlock Class}
\usage{
TransformerBlock(hidden, attn_heads, dropout)
}
\arguments{
\item{hidden}{Hidden size}

\item{attn_heads}{Number of attention heads}

\item{dropout}{Dropout rate}
}
\description{
Single Transformer encoder block with multi-head attention and FFN.
}
\keyword{internal}
