% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_TransformerModel.R
\name{TransformerLayer}
\alias{TransformerLayer}
\title{TransformerLayer Class}
\usage{
TransformerLayer(feature_size, heads = 1, dropout = 0.5, num_layers = 1)
}
\arguments{
\item{feature_size}{Hidden feature size}

\item{heads}{Number of attention heads}

\item{dropout}{Dropout rate}

\item{num_layers}{Number of transformer blocks}
}
\description{
Stack of Transformer encoder blocks.
}
\keyword{internal}
